Родолжаем делить модули
